+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.1.1 (Unreleased / YYYY-MM-DD)
 
+### Bugfixes
+- `ExtendNonAnyType` trait:  Fixed `array_buffer_view()` and `array_buffer_view_opt()` to properly return `NonAnyType::ArrayBufferView`. Previously, they returned `NonAnyType::ArrayBuffer`. ([#15](https://github.com/neoncitylights/webidl-utils/pull/15))
+
 ### Breaking changes
 - `ExtendNonAnyType` trait: the function signatures of `identifier()` and `identifier_opt()` are simplified. They now take a string literal. Previously, they both took an `Identifier` type (a newtype that stores a string literal) ([#12](https://github.com/neoncitylights/webidl-utils/pull/12))
 
@@ -9,6 +12,9 @@
 - Add `ExtendNonAnyType::is_required()` ([#14](https://github.com/neoncitylights/webidl-utils/pull/14))
 - Add `ExtendMayBeNull::is_required()` ([#13](https://github.com/neoncitylights/webidl-utils/pull/13))
 - Add and implement `ExtendRecordKeyType` trait for `weedle::types::RecordKeyType`
+
+## Internal changes
+- Expand code coverage for `ExtendNonAnyType` ([#15](https://github.com/neoncitylights/webidl-utils/pull/15))
 
 ## 0.1.0 (2024-06-27)